starter-chat-agents/local-llama-chat/requirements.txt
gradio>=4.0.0
llama-cpp-python>=0.2.0
python-dotenv>=1.0.0
